pyproject.toml
[tool.poetry]
name = "movies-service"
version = "0.1.0"
description = ""
authors = ["Artyom <nechepurenko.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
drf-spectacular = "^0.27.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
drf-extra-fields = "^3.7.0"
aiogram = "^3.17.0"
gunicorn = "^23.0.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
django-debug-toolbar = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
skip_glob = ["**/migrations/*.py"]
src_paths = ["src/backend/service", "src/bot"]
